Use newer X server's -displayfd option instead of random display number
With random display numbers there's still a chance that Xvfb would
try to use a display number that's already in use. The previous
implementation did not correctly detect this error case and just
continue leading to strange errors later.

The new implementation uses the -displayfd option that's available since
X.org 1.13 to let Xvfb chose an available display number and report it
back to WWW::WebKit. This also fixes the race condition between Xvfb
getting ready for connections and WWW::WebKit trying to access it.